features/articles/componentsの中にserverディレクトリを作成して、ArticlesPageTemplate.tsxを作成
